requirements.txt
flask==3.0.1
torch>=2.2.0
torchvision>=0.17.0
ultralytics 
transformers>=4.37.0
opencv-python>=4.9.0.80
pillow>=10.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
werkzeug>=3.0.1é

